+++ crates/stackable-versioned-macros/src/codegen/common/item.rs
@@ -1,7 +1,7 @@
 use std::{collections::BTreeMap, marker::PhantomData, ops::Deref};
 
 use quote::format_ident;
-use syn::{spanned::Spanned, Attribute, Ident, Path};
+use syn::{spanned::Spanned, Attribute, Ident, Path, Type};
 
 use crate::{
     attrs::common::{ContainerAttributes, ItemAttributes, ValidateVersions},
@@ -21,7 +21,7 @@ use crate::{
 pub(crate) trait Item<I, A>: Sized
 where
     A: for<'i> TryFrom<&'i I> + Attributes,
-    I: Named + Spanned,
+    I: InnerItem,
 {
     /// Creates a new versioned item (struct field or enum variant) by consuming
     /// the parsed [Field](syn::Field) or [Variant](syn::Variant) and validating
@@ -41,6 +41,10 @@ where
 
     /// Returns the ident of the item based on the provided container version.
     fn get_ident(&self, version: &ContainerVersion) -> Option<&Ident>;
+}
+
+pub(crate) trait InnerItem: Named + Spanned {
+    fn ty(&self) -> Type;
 }
 
 /// This trait enables access to the ident of named items, like fields and
@@ -87,11 +91,11 @@ pub(crate) trait Attributes {
 pub(crate) struct VersionedItem<I, A>
 where
     A: for<'i> TryFrom<&'i I> + Attributes,
-    I: Named + Spanned,
+    I: InnerItem,
 {
+    pub(crate) original_attributes: Vec<Attribute>,
     pub(crate) chain: Option<VersionChain>,
     pub(crate) inner: I,
-    pub(crate) original_attributes: Vec<Attribute>,
     _marker: PhantomData<A>,
 }
 
@@ -99,7 +103,7 @@ impl<I, A> Item<I, A> for VersionedItem<I, A>
 where
     syn::Error: for<'i> From<<A as TryFrom<&'i I>>::Error>,
     A: for<'i> TryFrom<&'i I> + Attributes + ValidateVersions<I>,
-    I: Named + Spanned,
+    I: InnerItem,
 {
     fn new(item: I, container_attrs: &ContainerAttributes) -> syn::Result<Self> {
         // We use the TryFrom trait here, because the type parameter `A` can use
@@ -125,37 +129,56 @@ where
 
         // Deprecating an item is always the last state an item can end up in.
         // For items which are not deprecated, the last change is either the
-        // latest rename or addition, which is handled below. The ident of the
+        // latest change or addition, which is handled below. The ident of the
         // deprecated item is guaranteed to include the 'deprecated_' or
         // 'DEPRECATED_' prefix. The ident can thus be used as is.
         if let Some(deprecated) = common_attributes.deprecated {
             let deprecated_ident = item.ident();
 
-            // When the item is deprecated, any rename which occurred beforehand
+            // When the item is deprecated, any change which occurred beforehand
             // requires access to the item ident to infer the item ident for
-            // the latest rename.
+            // the latest change.
             let mut ident = item.cleaned_ident();
+            let mut ty = item.ty();
+
             let mut actions = BTreeMap::new();
 
             actions.insert(
                 *deprecated.since,
-                ItemStatus::Deprecated {
+                ItemStatus::Deprecation {
                     previous_ident: ident.clone(),
                     ident: deprecated_ident.clone(),
-                    note: deprecated.note.to_string(),
+                    note: deprecated.note.as_deref().cloned(),
                 },
             );
 
-            for rename in common_attributes.renames.iter().rev() {
-                let from = format_ident!("{from}", from = *rename.from);
+            for change in common_attributes.changes.iter().rev() {
+                let from_ident = if let Some(from) = change.from_name.as_deref() {
+                    format_ident!("{from}")
+                } else {
+                    ident.clone()
+                };
+
+                // TODO (@Techassi): This is an awful lot of cloning, can we get
+                // rid of it?
+                let from_ty = change
+                    .from_type
+                    .as_ref()
+                    .map(|sv| sv.deref().clone())
+                    .unwrap_or(ty.clone());
+
                 actions.insert(
-                    *rename.since,
-                    ItemStatus::Renamed {
-                        from: from.clone(),
-                        to: ident,
+                    *change.since,
+                    ItemStatus::Change {
+                        from_ident: from_ident.clone(),
+                        to_ident: ident,
+                        from_type: from_ty.clone(),
+                        to_type: ty,
                     },
                 );
-                ident = from;
+
+                ident = from_ident;
+                ty = from_ty;
             }
 
             // After the last iteration above (if any) we use the ident for the
@@ -163,9 +186,10 @@ where
             if let Some(added) = common_attributes.added {
                 actions.insert(
                     *added.since,
-                    ItemStatus::Added {
+                    ItemStatus::Addition {
                         default_fn: added.default_fn.deref().clone(),
                         ident,
+                        ty,
                     },
                 );
             }
@@ -173,23 +197,42 @@ where
             Ok(Self {
                 _marker: PhantomData,
                 chain: Some(actions),
-                inner: item,
                 original_attributes,
+                inner: item,
             })
-        } else if !common_attributes.renames.is_empty() {
-            let mut actions = BTreeMap::new();
+        } else if !common_attributes.changes.is_empty() {
             let mut ident = item.ident().clone();
+            let mut ty = item.ty();
 
-            for rename in common_attributes.renames.iter().rev() {
-                let from = format_ident!("{from}", from = *rename.from);
+            let mut actions = BTreeMap::new();
+
+            for change in common_attributes.changes.iter().rev() {
+                let from_ident = if let Some(from) = change.from_name.as_deref() {
+                    format_ident!("{from}")
+                } else {
+                    ident.clone()
+                };
+
+                // TODO (@Techassi): This is an awful lot of cloning, can we get
+                // rid of it?
+                let from_ty = change
+                    .from_type
+                    .as_ref()
+                    .map(|sv| sv.deref().clone())
+                    .unwrap_or(ty.clone());
+
                 actions.insert(
-                    *rename.since,
-                    ItemStatus::Renamed {
-                        from: from.clone(),
-                        to: ident,
+                    *change.since,
+                    ItemStatus::Change {
+                        from_ident: from_ident.clone(),
+                        to_ident: ident,
+                        from_type: from_ty.clone(),
+                        to_type: ty,
                     },
                 );
-                ident = from;
+
+                ident = from_ident;
+                ty = from_ty;
             }
 
             // After the last iteration above (if any) we use the ident for the
@@ -197,9 +240,10 @@ where
             if let Some(added) = common_attributes.added {
                 actions.insert(
                     *added.since,
-                    ItemStatus::Added {
+                    ItemStatus::Addition {
                         default_fn: added.default_fn.deref().clone(),
                         ident,
+                        ty,
                     },
                 );
             }
@@ -207,8 +251,8 @@ where
             Ok(Self {
                 _marker: PhantomData,
                 chain: Some(actions),
-                inner: item,
                 original_attributes,
+                inner: item,
             })
         } else {
             if let Some(added) = common_attributes.added {
@@ -216,25 +260,26 @@ where
 
                 actions.insert(
                     *added.since,
-                    ItemStatus::Added {
+                    ItemStatus::Addition {
                         default_fn: added.default_fn.deref().clone(),
                         ident: item.ident().clone(),
+                        ty: item.ty(),
                     },
                 );
 
                 return Ok(Self {
                     _marker: PhantomData,
                     chain: Some(actions),
-                    inner: item,
                     original_attributes,
+                    inner: item,
                 });
             }
 
             Ok(Self {
                 _marker: PhantomData,
+                original_attributes,
                 chain: None,
                 inner: item,
-                original_attributes,
             })
         }
     }
@@ -248,39 +293,72 @@ where
 
                 match chain.get_neighbors(&version.inner) {
                     (None, Some(status)) => match status {
-                        ItemStatus::Added { .. } => {
+                        ItemStatus::Addition { .. } => {
                             chain.insert(version.inner, ItemStatus::NotPresent)
                         }
-                        ItemStatus::Renamed { from, .. } => {
-                            chain.insert(version.inner, ItemStatus::NoChange(from.clone()))
-                        }
-                        ItemStatus::Deprecated { previous_ident, .. } => chain
-                            .insert(version.inner, ItemStatus::NoChange(previous_ident.clone())),
-                        ItemStatus::NoChange(ident) => {
-                            chain.insert(version.inner, ItemStatus::NoChange(ident.clone()))
-                        }
+                        ItemStatus::Change {
+                            from_ident,
+                            from_type,
+                            ..
+                        } => chain.insert(
+                            version.inner,
+                            ItemStatus::NoChange {
+                                ident: from_ident.clone(),
+                                ty: from_type.clone(),
+                            },
+                        ),
+                        ItemStatus::Deprecation { previous_ident, .. } => chain.insert(
+                            version.inner,
+                            ItemStatus::NoChange {
+                                ident: previous_ident.clone(),
+                                ty: self.inner.ty(),
+                            },
+                        ),
+                        ItemStatus::NoChange { ident, ty } => chain.insert(
+                            version.inner,
+                            ItemStatus::NoChange {
+                                ident: ident.clone(),
+                                ty: ty.clone(),
+                            },
+                        ),
                         ItemStatus::NotPresent => unreachable!(),
                     },
                     (Some(status), None) => {
-                        let ident = match status {
-                            ItemStatus::Added { ident, .. } => ident,
-                            ItemStatus::Renamed { to, .. } => to,
-                            ItemStatus::Deprecated { ident, .. } => ident,
-                            ItemStatus::NoChange(ident) => ident,
+                        let (ident, ty) = match status {
+                            ItemStatus::Addition { ident, ty, .. } => (ident, ty),
+                            ItemStatus::Change {
+                                to_ident, to_type, ..
+                            } => (to_ident, to_type),
+                            ItemStatus::Deprecation { ident, .. } => (ident, &self.inner.ty()),
+                            ItemStatus::NoChange { ident, ty } => (ident, ty),
                             ItemStatus::NotPresent => unreachable!(),
                         };
 
-                        chain.insert(version.inner, ItemStatus::NoChange(ident.clone()))
+                        chain.insert(
+                            version.inner,
+                            ItemStatus::NoChange {
+                                ident: ident.clone(),
+                                ty: ty.clone(),
+                            },
+                        )
                     }
                     (Some(status), Some(_)) => {
-                        let ident = match status {
-                            ItemStatus::Added { ident, .. } => ident,
-                            ItemStatus::Renamed { to, .. } => to,
-                            ItemStatus::NoChange(ident) => ident,
+                        let (ident, ty) = match status {
+                            ItemStatus::Addition { ident, ty, .. } => (ident, ty),
+                            ItemStatus::Change {
+                                to_ident, to_type, ..
+                            } => (to_ident, to_type),
+                            ItemStatus::NoChange { ident, ty, .. } => (ident, ty),
                             _ => unreachable!(),
                         };
 
-                        chain.insert(version.inner, ItemStatus::NoChange(ident.clone()))
+                        chain.insert(
+                            version.inner,
+                            ItemStatus::NoChange {
+                                ident: ident.clone(),
+                                ty: ty.clone(),
+                            },
+                        )
                     }
                     _ => unreachable!(),
                 };
@@ -299,32 +377,40 @@ where
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub(crate) enum ItemStatus {
-    Added {
+    Addition {
         ident: Ident,
         default_fn: Path,
+        // NOTE (@Techassi): We need to carry idents and type information in
+        // nearly every status. Ideally, we would store this in separate maps.
+        ty: Type,
     },
-    Renamed {
-        from: Ident,
-        to: Ident,
+    Change {
+        from_ident: Ident,
+        to_ident: Ident,
+        from_type: Type,
+        to_type: Type,
     },
-    Deprecated {
+    Deprecation {
         previous_ident: Ident,
+        note: Option<String>,
         ident: Ident,
-        note: String,
     },
-    NoChange(Ident),
+    NoChange {
+        ident: Ident,
+        ty: Type,
+    },
     NotPresent,
 }
 
 impl ItemStatus {
     pub(crate) fn get_ident(&self) -> Option<&Ident> {
         match &self {
-            ItemStatus::Added { ident, .. } => Some(ident),
-            ItemStatus::Renamed { to, .. } => Some(to),
-            ItemStatus::Deprecated { ident, .. } => Some(ident),
-            ItemStatus::NoChange(ident) => Some(ident),
+            ItemStatus::Addition { ident, .. } => Some(ident),
+            ItemStatus::Change { to_ident, .. } => Some(to_ident),
+            ItemStatus::Deprecation { ident, .. } => Some(ident),
+            ItemStatus::NoChange { ident, .. } => Some(ident),
             ItemStatus::NotPresent => None,
         }
     }