+++ crates/stackable-versioned-macros/src/codegen/vstruct/field.rs
@@ -11,8 +11,8 @@ use crate::{
         field::FieldAttributes,
     },
     codegen::common::{
-        remove_deprecated_field_prefix, Attributes, ContainerVersion, Item, ItemStatus, Named,
-        VersionedItem,
+        remove_deprecated_field_prefix, Attributes, ContainerVersion, InnerItem, Item, ItemStatus,
+        Named, VersionedItem,
     },
 };
 
@@ -60,6 +60,12 @@ impl Attributes for FieldAttributes {
 
     fn original_attributes(&self) -> &Vec<syn::Attribute> {
         &self.attrs
+    }
+}
+
+impl InnerItem for Field {
+    fn ty(&self) -> syn::Type {
+        self.ty.clone()
     }
 }
 
@@ -115,27 +121,45 @@ impl VersionedField {
                         container_version.inner
                     )
                 }) {
-                    ItemStatus::Added { ident, .. } => Some(quote! {
+                    ItemStatus::Addition { ident, ty, .. } => Some(quote! {
                         #(#original_attributes)*
-                        pub #ident: #field_type,
+                        pub #ident: #ty,
                     }),
-                    ItemStatus::Renamed { to, .. } => Some(quote! {
+                    ItemStatus::Change {
+                        to_ident, to_type, ..
+                    } => Some(quote! {
                         #(#original_attributes)*
-                        pub #to: #field_type,
+                        pub #to_ident: #to_type,
                     }),
-                    ItemStatus::Deprecated {
+                    ItemStatus::Deprecation {
                         ident: field_ident,
                         note,
                         ..
-                    } => Some(quote! {
-                        #(#original_attributes)*
-                        #[deprecated = #note]
-                        pub #field_ident: #field_type,
-                    }),
+                    } => {
+                        // FIXME (@Techassi): Emitting the deprecated attribute
+                        // should cary over even when the item status is
+                        // 'NoChange'.
+                        // TODO (@Techassi): Make the generation of deprecated
+                        // items customizable. When a container is used as a K8s
+                        // CRD, the item must continue to exist, even when
+                        // deprecated. For other versioning use-cases, that
+                        // might not be the case.
+                        let deprecated_attr = if let Some(note) = note {
+                            quote! {#[deprecated = #note]}
+                        } else {
+                            quote! {#[deprecated]}
+                        };
+
+                        Some(quote! {
+                            #(#original_attributes)*
+                            #deprecated_attr
+                            pub #field_ident: #field_type,
+                        })
+                    }
                     ItemStatus::NotPresent => None,
-                    ItemStatus::NoChange(field_ident) => Some(quote! {
+                    ItemStatus::NoChange { ident, ty } => Some(quote! {
                         #(#original_attributes)*
-                        pub #field_ident: #field_type,
+                        pub #ident: #ty,
                     }),
                 }
             }
@@ -170,9 +194,33 @@ impl VersionedField {
                         .get(&next_version.inner)
                         .expect("internal error: chain must contain container version"),
                 ) {
-                    (_, ItemStatus::Added { ident, default_fn }) => quote! {
+                    (
+                        _,
+                        ItemStatus::Addition {
+                            ident, default_fn, ..
+                        },
+                    ) => quote! {
                         #ident: #default_fn(),
                     },
+                    (
+                        _,
+                        ItemStatus::Change {
+                            from_ident: old_field_ident,
+                            to_ident,
+                            from_type,
+                            to_type,
+                        },
+                    ) => {
+                        if from_type == to_type {
+                            quote! {
+                                #to_ident: #from_ident.#old_field_ident,
+                            }
+                        } else {
+                            quote! {
+                                #to_ident: #from_ident.#old_field_ident.into(),
+                            }
+                        }
+                    }
                     (old, next) => {
                         let old_field_ident = old
                             .get_ident()