+++ crates/stackable-versioned-macros/tests/default/pass/basic.rs
@@ -11,10 +11,10 @@ use stackable_versioned_macros::versioned;
     version(name = "v2"),
     version(name = "v3")
 )]
-struct Foo {
+pub(crate) struct Foo {
     #[versioned(
-        added(since = "v1alpha1"),
-        renamed(since = "v1beta1", from = "jjj"),
+        changed(since = "v1beta1", from_name = "jjj", from_type = "u8"),
+        changed(since = "v1", from_type = "u16"),
         deprecated(since = "v2", note = "not empty")
     )]
     /// Test